[Entity] Improve physics activation logic
Enabling physics will generate a default box body around the object. This only happens if there isn't any body yet created for the entity.

It now also stores a reference for the scene it's currently applied to, so it can now add itself to the scene physical world if the physics bodies were created after the entity was added to the scene